Prevent call to function on non-object in ping
When wrapping an existing PDO instance using the `'pdo'` param, `$this->platform` is not set because the `connect` method is bypassed. Calling `getDatabasePlatform()` explicitly ensures the platform is correctly set, or at least that auto-detection is attempted.

Not sure if this is the best way to fix this but it was enough to get it to work for me.